move difficulty tabs back to `modsTable.tsx`
