+++ src/components/mods/modsTable.tsx
@@ -1,5 +1,6 @@
 import { DataTable, DataTableSortStatus } from "mantine-datatable";
 import { Dispatch, SetStateAction, useEffect, useMemo, useRef, useState, createContext } from "react";
+import { createPortal } from "react-dom";
 import { ExpandedMod } from "~/components/mods/expandedMod";
 import { createStyles, Text } from "@mantine/core";
 import { useDebouncedValue } from "@mantine/hooks";
@@ -14,11 +15,11 @@ import type { ModWithInfo, Tech } from "~/components/mods/types";
 import { noRatingsFoundMessage } from "~/consts/noRatingsFoundMessage";
 import { defaultToLocaleDateStringOptions } from "~/consts/defaultToLocaleDateStringOptions";
 import { colorsForDifficultyIndex, greatestValidDifficultyIndex } from "~/styles/modsColors";
-import { canonicalDifficultyNames, type DifficultyColor } from "~/styles/difficultyColors";
+import { canonicalDifficultyNames, difficultyColors, type DifficultyColor } from "~/styles/difficultyColors";
 import { expandedModColors } from "~/styles/expandedModColors";
 import { TABLE_HEADER_ARROW_ZOOM } from "~/consts/tableHeaderArrowZoom";
 import { pageContentHeightPixels } from "~/styles/pageContentHeightPixels";
-import { DifficultyTabs } from "./difficultyTabs";
+import { blackBackgroundColor } from "~/styles/layoutColors";
 
 
 
@@ -28,6 +29,7 @@ export const currentDifficultyTabIndexContext = createContext<number | null>(nul
 const PAGE_SIZES = [5, 10, 15, 20, 25, 50, 100, 250, 500, 1000];
 const DEFAULT_PAGE_SIZE_INDEX = 1;
 const QUERY_DEBOUNCE_TIME_MILLISECONDS = 200;
+const ACTIVE_DIFFICULTY_TAB_BORDER_HEIGHT = "2px";
 
 // TODO: remove these parameters, limit the width of the columns in the table in some way, and let the datatable columns' `ellipsis` property handle the overflow.
 const NAME_COLUMN_MAX_LETTERS = 35;
@@ -45,6 +47,87 @@ const useStyles = createStyles(
         { colors }: { colors: DifficultyColor; },
     ) => {
         return ({
+            tabContainer: {
+                position: "sticky",
+                top: "0",
+                zIndex: 2,
+                display: "flex",
+                justifyContent: "end",
+                padding: "0 15px",
+                backgroundColor: blackBackgroundColor,
+            },
+            tab: {
+                padding: "1px 20px",
+                display: "inline-block",
+                borderTopLeftRadius: "5px",
+                borderTopRightRadius: "5px",
+                borderTop: ACTIVE_DIFFICULTY_TAB_BORDER_HEIGHT,
+                borderTopStyle: "solid",
+                borderTopColor: "transparent",
+                fontSize: "medium",
+                cursor: "pointer",
+                fontWeight: "bold",
+            },
+            activeTab: {
+                borderTopColor: `${colors.primaryHover.backgroundColor}`,  // add top border to active tab so it's easier to see which tab is active (the contrast ratio between the difficulty colors is not sufficient on its own)
+            },
+            // color the difficulty tabs
+            beginner: {
+                backgroundColor: difficultyColors.beginner.primary.backgroundColor,
+                color: difficultyColors.beginner.primary.textColor,
+                ":hover": {
+                    backgroundColor: difficultyColors.beginner.primaryHover.backgroundColor,
+                    color: difficultyColors.beginner.primaryHover.textColor,
+                },
+            },
+            intermediate: {
+                backgroundColor: difficultyColors.intermediate.primary.backgroundColor,
+                color: difficultyColors.intermediate.primary.textColor,
+                ":hover": {
+                    backgroundColor: difficultyColors.intermediate.primaryHover.backgroundColor,
+                    color: difficultyColors.intermediate.primaryHover.textColor,
+                },
+            },
+            advanced: {
+                backgroundColor: difficultyColors.advanced.primary.backgroundColor,
+                color: difficultyColors.advanced.primary.textColor,
+                ":hover": {
+                    backgroundColor: difficultyColors.advanced.primaryHover.backgroundColor,
+                    color: difficultyColors.advanced.primaryHover.textColor,
+                },
+            },
+            expert: {
+                backgroundColor: difficultyColors.expert.primary.backgroundColor,
+                color: difficultyColors.expert.primary.textColor,
+                ":hover": {
+                    backgroundColor: difficultyColors.expert.primaryHover.backgroundColor,
+                    color: difficultyColors.expert.primaryHover.textColor,
+                },
+            },
+            grandmaster: {
+                backgroundColor: difficultyColors.grandmaster.primary.backgroundColor,
+                color: difficultyColors.grandmaster.primary.textColor,
+                ":hover": {
+                    backgroundColor: difficultyColors.grandmaster.primaryHover.backgroundColor,
+                    color: difficultyColors.grandmaster.primaryHover.textColor,
+                },
+            },
+            astral: {
+                backgroundColor: difficultyColors.astral.primary.backgroundColor,
+                color: difficultyColors.astral.primary.textColor,
+                ":hover": {
+                    backgroundColor: difficultyColors.astral.primaryHover.backgroundColor,
+                    color: difficultyColors.astral.primaryHover.textColor,
+                },
+            },
+            celestial: {
+                backgroundColor: difficultyColors.celestial.primary.backgroundColor,
+                color: difficultyColors.celestial.primary.textColor,
+                ":hover": {
+                    backgroundColor: difficultyColors.celestial.primaryHover.backgroundColor,
+                    color: difficultyColors.celestial.primaryHover.textColor,
+                },
+            },
             table: {
                 "&&&& table": {
                     transform: "translate(0, -21px)",
@@ -681,12 +764,6 @@ export const ModsTable = ({ qualities, difficulties, techs, modsWithInfo, isLoad
 
 
 
-    // We want to render the tab container inside the scroll area of the datatable, so we use ref and portal.
-    const tableBodyRef = useRef<HTMLTableSectionElement>(null);
-
-
-
-
     // apply the correct class (defined in ~/styles/globals.css) to the body element to change the background color of the pagination dropdown
     useEffect(() => {
         const menuClassNames = canonicalDifficultyNames.map((difficultyName) => `${difficultyName.toLowerCase()}-menu`);
@@ -723,15 +800,76 @@ export const ModsTable = ({ qualities, difficulties, techs, modsWithInfo, isLoad
     const { cx, classes } = useStyles({ colors });
 
 
+    const tabColors: string[] = Array(canonicalDifficultyNames.length);
+
+    Object.entries(classes).forEach(
+        ([key, value]) => {
+            for (let index = 0; index < canonicalDifficultyNames.length; index++) {
+                if (key === canonicalDifficultyNames[index]) {
+                    tabColors[index] = value;
+                    break;
+                }
+            }
+        }
+    );
+
+
+    // We want to render the tab container inside the scroll area of the datatable, so we use ref and portal.
+    const tableBodyRef = useRef<HTMLTableSectionElement>(null);
+    const [tabContainer, setTabContainer] = useState<HTMLDivElement | null>(null);
+
+    useEffect(() => {
+        const tabsParent = tableBodyRef.current?.parentElement?.parentElement;
+
+        if (!tabsParent) {
+            throw "Couldn't find tabsParent.";
+        }
+
+
+        const tabContainer = document.createElement("div");
+
+        tabContainer.className = classes.tabContainer;
+
+        tabsParent.prepend(tabContainer);
+
+        setTabContainer(tabContainer);
+
+
+        return () => {
+            tabsParent.removeChild(tabContainer);
+            
+            setTabContainer(null);
+        };
+    }, [classes.tabContainer]);
+
+
     return (
         <>
-            <DifficultyTabs
-                colors={colors}
-                tableBodyRef={tableBodyRef}
-                parentDifficultyNames={parentDifficultyNames}
-                currentTabIndex={currentTabIndex}
-                setCurrentTabIndex={setCurrentTabIndex}
-            />
+            {
+                tabContainer !== null && (
+                    createPortal(
+                        [...parentDifficultyNames].reverse().map(
+                            (name, index) =>
+                                <span
+                                    key={name}
+                                    className={
+                                        cx(
+                                            classes.tab,
+                                            tabColors[parentDifficultyNames.length - 1 - index],
+                                            { [classes.activeTab]: parentDifficultyNames.length - 1 - index === currentTabIndex }
+                                        )
+                                    }
+                                    onClick={() => {
+                                        setCurrentTabIndex(parentDifficultyNames.length - 1 - index);
+                                    }}
+                                >
+                                    {name}
+                                </span>
+                        ),
+                        tabContainer
+                    )
+                )
+            }
             <currentDifficultyTabIndexContext.Provider value={currentTabIndex}>
                 <DataTable
                     bodyRef={tableBodyRef}
